+++ bee-storage/bee-storage-null/src/access/exist.rs
@@ -1,0 +1,12 @@
+// Copyright 2021 IOTA Stiftung
+// SPDX-License-Identifier: Apache-2.0
+
+use crate::Storage;
+
+use bee_storage::access::Exist;
+
+impl<K, V> Exist<K, V> for Storage {
+    fn exist(&self, _key: &K) -> Result<bool, Self::Error> {
+        Ok(false)
+    }
+}
