+++ bee-storage/bee-storage-null/src/access/mod.rs
@@ -2,5 +2,10 @@
 // SPDX-License-Identifier: Apache-2.0
 
 pub mod batch;
+pub mod delete;
+pub mod exist;
+pub mod fetch;
+pub mod insert;
 pub mod iter;
 pub mod multi_fetch;
+pub mod truncate;
