+++ bee-storage/bee-storage-null/src/lib.rs
@@ -3,10 +3,7 @@
 
 pub mod access;
 
-use bee_storage::{
-    access::{Delete, Exist, Fetch, Insert, Truncate},
-    backend::StorageBackend,
-};
+use bee_storage::backend::StorageBackend;
 
 pub struct Storage;
 
@@ -32,36 +29,6 @@ impl StorageBackend for Storage {
     }
 
     fn set_health(&self, _health: bee_storage::system::StorageHealth) -> Result<(), Self::Error> {
-        Ok(())
-    }
-}
-
-impl<K, V> Delete<K, V> for Storage {
-    fn delete(&self, _key: &K) -> Result<(), Self::Error> {
-        Ok(())
-    }
-}
-
-impl<K, V> Exist<K, V> for Storage {
-    fn exist(&self, _key: &K) -> Result<bool, Self::Error> {
-        Ok(false)
-    }
-}
-
-impl<K, V> Fetch<K, V> for Storage {
-    fn fetch(&self, _key: &K) -> Result<Option<V>, Self::Error> {
-        Ok(None)
-    }
-}
-
-impl<K, V> Insert<K, V> for Storage {
-    fn insert(&self, _key: &K, _value: &V) -> Result<(), Self::Error> {
-        Ok(())
-    }
-}
-
-impl<K, V> Truncate<K, V> for Storage {
-    fn truncate(&self) -> Result<(), Self::Error> {
         Ok(())
     }
 }
